requirements.txt
pycryptodomex>=3.4.7
aiocoap=0.4.1